CLEANING
Fixes after merging with testing

arc handling w/svgparser mostly working, added rect line thickness awareness

Split out a separate application (somewhat poorly) for level previewing. Started on basic camera controls. Currently loads level every frame--need to implement file watcher. Added CSS color declaration parser for futureproofing vs just parsing hex web colors.

started setting up loading svgs dynamically

add pict2svg to makefile

fixed self-intersecting faces, should also be double-sided faces

a first attempt at proper lighting

Fix syntax error in shader

refactor geometry shader to be a little more readable and remove AvaraGL stuff from CViewParameters

attempt to limit output to 16 bit color

remove unnecessary include in svg parser

normals are being rotated but who knows if they are right

fix a few things in pict2svg.py

commiting WIP rendering system update as laptop battery dies -- moved transformation steps into vertex shader and passing in several matrices, doesn't work yet

closer but the Y axis of the final image seems to be reversed

moved projection matrix out of vertex shader and figured out x/z flipping. now there is a new bug with vertex data and lots of opengl 502 errors

successfully refactored to only upload vertex data a few times (at instantiation and then for every vertex change like scaling) instead of every frame

trying to fix rendering bugs

Bug caused by incorrect shader program being selected during uniform transfer was fixed, now working with new shader based on transforms and double sided geometry. shading still has some issues and it has introduced more z-fighting for objects close to the camera

remove extra includes from AvaraGL

fix windows travis build

Remove execinfo.h -- no longer needed

refactor projection matrix to be modified at runtime for different window sizes--default fov 45

sky orientation fixed with skybox geometry but it has seams at the corners

fixed sky shader

Take us back to single sided shapes with FACE_CULL off. We will attempt to reverse normals on the fly in the shader. Trying to figure out a baseline for normal transform--the current normal transform is not working correctly, but shape normals seem to be mostly accurate

Much brighter--normal transform applied, using gl_FrontFacing in fragment shader to shade bottom of single side geometry correctly

Add 'decal' support to the shader--lifts flat geometry in the z-buffer slightly to counteract z-fighting

clearer z-value scaling for decals

reverse entire normal, not just y value, if not front face

lighting looks much closer to original but not sure of the lights actual direction matching

Another pass on working lighting--internally consistent but doesn't match original

Add dan's patch to fix the lighting!

clean up branch for merge, move cartesian conversion of light vectors out of fragment shader

clean up matrix conversion from fixed to float

+++ src/bsp/CBSPWorld.cpp
@@ -260,7 +260,7 @@ void CBSPWorldImpl::SortVisibleParts() {
 }
 
 void CBSPWorldImpl::Render(CViewParameters *theView) {
-    short i;
+    int i;
     CBSPPart **sp, **sd;
 
     currentView = theView;
